fix: Phase 1 compilation blockers - systematic architectural fixes
Fixed 817 compiler errors identified by comprehensive code quality analysis:
- Added missing types: ScreeningMode and ScreeningStatus enums
- Completed interface implementations: GetRealTimeDataAsync, LogDebug methods
- Fixed exception scope issues where 'ex' was referenced outside catch blocks
- Implemented ValidateCriteria method with proper business rule validation
- Added all 5 missing IProcessManager interface methods for Windows optimization
- Fixed MarketData constructor to use ITradingLogger parameter correctly

Systematic fixes following holistic architecture principles, maintaining
interface segregation and dependency inversion throughout.

+++ DayTradinPlatform/TradingPlatform.Core/Logging/PerformanceMonitor.cs
@@ -89,7 +89,7 @@ internal class PerformanceMonitor
             return new PerformanceContext
             {
                 Operation = methodName,
-                DurationNanoseconds = stats.TotalDurationNs / Math.Max(1, stats.CallCount),
+                DurationNanoseconds = (long)(stats.TotalDurationNs / Math.Max(1, stats.CallCount)),
                 DurationMilliseconds = (stats.TotalDurationNs / Math.Max(1, stats.CallCount)) / 1_000_000.0,
                 ResourceUsage = new Dictionary<string, object>
                 {
